Extract pdf generation and move handling into dedicated services (#760)
- Introduces `pdf_generation_service` to handle PDF rendering,
  previews, filename computation, and language/runtime config.
- Introduces `item_move_service` to
  encapsulate page/element move/resequence logic.
